Small cleanup...

+++ External/Plugins/ASCompletion/Completion/ASComplete.cs
@@ -418,9 +418,10 @@ namespace ASCompletion.Completion
         {
             if (!ASContext.CommonSettings.AddClosingBraces) return;
             var context = ASContext.Context;
+            var currentPos = sci.CurrentPos;
             if (addedChar)
             {
-                if (IsMatchingQuote(c, sci.BaseStyleAt(sci.CurrentPos - 2)) && context.CodeComplete.IsEscapedCharacter(sci, sci.CurrentPos - 1))
+                if (IsMatchingQuote(c, sci.BaseStyleAt(currentPos - 2)) && context.CodeComplete.IsEscapedCharacter(sci, currentPos - 1))
                 {
                     return;
                 }
@@ -436,24 +437,24 @@ namespace ASCompletion.Completion
                     // Get the before & after style values unaffected by the entered char
                     sci.DeleteBack();
                     sci.Colourise(0, -1);
-                    styleBefore = (byte) sci.BaseStyleAt(sci.CurrentPos - 1);
-                    styleAfter = (byte) sci.BaseStyleAt(sci.CurrentPos);
+                    styleBefore = (byte) sci.BaseStyleAt(currentPos - 1);
+                    styleAfter = (byte) sci.BaseStyleAt(currentPos);
                     sci.AddText(1, c.ToString());
                     undo = true;
                 }
                 else
                 {
-                    styleBefore = (byte) sci.BaseStyleAt(sci.CurrentPos - 2);
-                    styleAfter = (byte) sci.BaseStyleAt(sci.CurrentPos);
+                    styleBefore = (byte) sci.BaseStyleAt(currentPos - 2);
+                    styleAfter = (byte) sci.BaseStyleAt(currentPos);
                 }
 
                 // not inside a string literal
-                int position = sci.CurrentPos - 1;
+                int position = currentPos - 1;
                 if (!(IsStringStyle(styleBefore) && IsStringStyle(styleAfter)) && !(IsCharStyle(styleBefore) && IsCharStyle(styleAfter))
                     || context.CodeComplete.IsStringInterpolationStyle(sci, position))
                 {
                     char nextChar = sci.CurrentChar;
-                    int nextPos = sci.CurrentPos;
+                    int nextPos = currentPos;
 
                     while (nextChar == ' ' || nextChar == '\t') // Don't skip new line characters
                     {
@@ -475,7 +476,7 @@ namespace ASCompletion.Completion
                 else if (IsMatchingQuote(c, styleAfter))
                 {
                     char nextChar = sci.CurrentChar;
-                    int nextPos = sci.CurrentPos;
+                    int nextPos = currentPos;
 
                     while (nextChar == ' ' || nextChar == '\t') // Don't skip new line characters
                     {
@@ -498,23 +499,23 @@ namespace ASCompletion.Completion
             }
             else
             {
-                char open = (char) sci.CharAt(sci.CurrentPos - 1);
+                char open = (char) sci.CharAt(currentPos - 1);
 
-                if (IsMatchingQuote(open, sci.BaseStyleAt(sci.CurrentPos - 2)) && context.CodeComplete.IsEscapedCharacter(sci, sci.CurrentPos - 1))
+                if (IsMatchingQuote(open, sci.BaseStyleAt(currentPos - 2)) && context.CodeComplete.IsEscapedCharacter(sci, currentPos - 1))
                 {
                     return;
                 }
 
-                int styleBefore = sci.BaseStyleAt(sci.CurrentPos - 2);
-                int styleAfter = sci.BaseStyleAt(sci.CurrentPos);
+                int styleBefore = sci.BaseStyleAt(currentPos - 2);
+                int styleAfter = sci.BaseStyleAt(currentPos);
 
                 // not inside a string literal
-                int position = sci.CurrentPos - 1;
+                int position = currentPos - 1;
                 if (!(IsStringStyle(styleBefore) && IsStringStyle(styleAfter)) && !(IsCharStyle(styleBefore) && IsCharStyle(styleAfter))
                     || context.CodeComplete.IsStringInterpolationStyle(sci, position)
                     || IsMatchingQuote(open, styleAfter))
                 {
-                    int closePos = sci.CurrentPos;
+                    int closePos = currentPos;
 
                     while (char.IsWhiteSpace(c))
                     {
@@ -556,15 +557,12 @@ namespace ASCompletion.Completion
 
         private static bool HandleBraceClose(ScintillaControl sci, Brace brace, char close, char next, int nextPosition)
         {
-            if (close == brace.Close && next == brace.Close)
+            if (close == brace.Close && next == brace.Close && brace.ShouldClose(sci.CurrentPos, nextPosition))
             {
-                if (brace.ShouldClose(sci.CurrentPos, nextPosition))
-                {
-                    sci.DeleteBack();
-                    sci.AnchorPosition = nextPosition;
-                    sci.CurrentPos = nextPosition;
-                    return true;
-                }
+                sci.DeleteBack();
+                sci.AnchorPosition = nextPosition;
+                sci.CurrentPos = nextPosition;
+                return true;
             }
 
             return false;
@@ -572,14 +570,11 @@ namespace ASCompletion.Completion
         
         private static bool HandleBraceRemove(ScintillaControl sci, Brace brace, char open, char close, int closePosition)
         {
-            if (open == brace.Open && close == brace.Close)
+            if (open == brace.Open && close == brace.Close && brace.ShouldRemove(sci.CurrentPos, closePosition))
             {
-                if (brace.ShouldRemove(sci.CurrentPos, closePosition))
-                {
-                    sci.SelectionEnd = closePosition + 1;
-                    sci.DeleteBack();
-                    return true;
-                }
+                sci.SelectionEnd = closePosition + 1;
+                sci.DeleteBack();
+                return true;
             }
 
             return false;
